+++ R/plot.gllvm.R
@@ -6,7 +6,7 @@
 #' @param which if a subset of the plots is required, specify a subset of the numbers 1:5, see caption below.
 #' @param caption captions to appear above the plots.
 #' @param var.colors colors for responses, vector with length of number of response variables or 1. Defaults to NULL, when different responses have different colors.
-#' @param add.smooth	logical indicating if a smoother should be added.
+#' @param add.smooth logical with default \code{TRUE}. Indicates if a smoother should be added.
 #' @param envelopes logical, indicating if simulated point-wise confidence interval envelope will be added to Q-Q plot, defaults to \code{TRUE}
 #' @param reps number of replications when simulating confidence envelopes for normal Q-Q plot
 #' @param envelope.col colors for envelopes, vector with length of two
@@ -44,6 +44,11 @@ plot.gllvm <- function(x, which = 1:5, caption = c("Residuals vs linear predicto
                                                    "Residuals vs row", "Residuals vs column", "Scale-Location"), 
                        var.colors = NULL, add.smooth = TRUE, envelopes = TRUE, reps = 150, 
                        envelope.col = c("blue","lightblue"), n.plot = NULL, ...) {
+  
+  if(missing(caption)) caption <- c("Residuals vs linear predictors", "Normal Q-Q",
+                                    "Residuals vs row", "Residuals vs column", "Scale-Location")[which]
+  if(length(caption)!=length(which))stop("'caption' should have the same length as 'which'.")
+  
   n <- NROW(x$y)
   p <- NCOL(x$y)
   
@@ -52,10 +57,7 @@ plot.gllvm <- function(x, which = 1:5, caption = c("Residuals vs linear predicto
     sppind <- sort(sample(1:p, n.plot))
     p <- n.plot
   }
-  
-  mains <- rep("", 4)
-  mains[which] <- caption[which]
-  
+
   res <- residuals(x)
   ds.res <- res$residuals[,sppind]
 
@@ -78,17 +80,17 @@ plot.gllvm <- function(x, which = 1:5, caption = c("Residuals vs linear predicto
   if(1 %in% which) {
     if(is.null(gr.pars$xlim)) {
       plot(eta.mat[!is.na(ds.res)], ds.res[!is.na(ds.res)], xlab = "linear predictors", ylab = "Dunn-Smyth-residuals",
-           type = "n", col = rep(col[csum], each = n)[!is.na(ds.res)], main = mains[1], xlim = c(min(xxx), max(xxx))); abline(0, 0, col = "grey", lty = 3)
+           type = "n", col = rep(col[csum], each = n)[!is.na(ds.res)], main = caption[which==1], xlim = c(min(xxx), max(xxx))); abline(0, 0, col = "grey", lty = 3)
     } else {
       plot(eta.mat[!is.na(ds.res)], ds.res[!is.na(ds.res)], xlab = "linear predictors", ylab = "Dunn-Smyth-residuals", type =
-             "n", col = rep(col[csum], each = n), main = mains[1], ...); abline(0, 0, col = "grey", lty = 3)
+             "n", col = rep(col[csum], each = n), main = caption[which==1], ...); abline(0, 0, col = "grey", lty = 3)
     }
     
     if(add.smooth) gamEnvelope(eta.mat[!is.na(ds.res)], ds.res[!is.na(ds.res)], col = rep(col[csum], each = n)[!is.na(ds.res)], envelopes = envelopes, envelope.col = envelope.col, ...)
     #      panel(eta.mat, ds.res, col = rep(col, each = n), cex = 1, cex.lab = 1, cex.axis = 1, lwd = 1)
   }
   if(2 %in% which) {
-    qq.x<-qqnorm(c(ds.res)[!is.na(ds.res)], main = mains[2], ylab = "Dunn-Smyth residuals", col = rep(col[csum], each = n)[!is.na(ds.res)], cex = 0.5, xlab = "theoretical quantiles");
+    qq.x<-qqnorm(c(ds.res)[!is.na(ds.res)], main = caption[which==2], ylab = "Dunn-Smyth residuals", col = rep(col[csum], each = n)[!is.na(ds.res)], cex = 0.5, xlab = "theoretical quantiles");
     qqline(c(res$residuals)[!is.na(ds.res)], col = envelope.col[1])
     if(envelopes){
       K <- reps
@@ -112,14 +114,14 @@ plot.gllvm <- function(x, which = 1:5, caption = c("Residuals vs linear predicto
   }
   if(3 %in% which) {
     plot(rep(1:n, p)[!is.na(ds.res)], ds.res[!is.na(ds.res)], xlab = "site index", ylab = "Dunn-Smyth-residuals", col =
-           rep(col[csum], each = n)[!is.na(ds.res)], main = mains[3], ...);
+           rep(col[csum], each = n)[!is.na(ds.res)], main = caption[which==3], ...);
     abline(0, 0, col = "grey", lty = 3)
     if(add.smooth) panel.smooth(rep(1:n, p)[!is.na(ds.res)], ds.res[!is.na(ds.res)], col = rep(col[csum], each = n)[!is.na(ds.res)], col.smooth = envelope.col[1], cex = NA, ...)
     #panel(rep(1:n, p), ds.res, col = rep(col, each = n), cex = 1, cex.lab = 1, cex.axis = 1, lwd = 1)
   }
   if(4 %in% which) {
     plot(rep(1:p, each = n)[!is.na(ds.res)], ds.res[!is.na(ds.res)], xlab = "species index", ylab = "Dunn-Smyth-residuals", col =
-           rep(col[csum], each = n)[!is.na(ds.res)], main = mains[4], ...);  abline(0, 0, col = "grey", lty = 3)
+           rep(col[csum], each = n)[!is.na(ds.res)], main = caption[which==4], ...);  abline(0, 0, col = "grey", lty = 3)
     if(add.smooth) panel.smooth(rep(1:p, each = n)[!is.na(ds.res)], ds.res[!is.na(ds.res)], col = rep(col[csum], each = n)[!is.na(ds.res)], col.smooth = envelope.col[1], cex = NA, ...)
     #panel(rep(1:p, each = n), ds.res, col = rep(col[csum], each = n), cex = 1, cex.lab = 1, cex.axis = 1, lwd = 1)
   }
@@ -128,9 +130,9 @@ plot.gllvm <- function(x, which = 1:5, caption = c("Residuals vs linear predicto
     yl <- as.expression(substitute(sqrt(abs(YL)), list(YL = as.name("Dunn-Smyth-residuals"))))
     if(is.null(gr.pars$xlim)) {
       plot(eta.mat[!is.na(ds.res)], sqres, xlab = "linear predictors", ylab = yl, col = rep(col[csum], each = n)[!is.na(ds.res)],
-           main = mains[5], xlim = c(min(xxx), max(xxx)), ...);
+           main = caption[which==5], xlim = c(min(xxx), max(xxx)), ...);
     } else {
-      plot(eta.mat[!is.na(ds.res)], sqres, xlab = "linear predictors", ylab = yl, col = rep(col[csum], each = n)[!is.na(ds.res)], main = mains[5], ...);
+      plot(eta.mat[!is.na(ds.res)], sqres, xlab = "linear predictors", ylab = yl, col = rep(col[csum], each = n)[!is.na(ds.res)], main = caption[which==5], ...);
     }
     if(add.smooth) panel.smooth(eta.mat[!is.na(ds.res)], sqres, col = rep(col[csum], each = n)[!is.na(ds.res)], col.smooth = envelope.col[1], cex = NA, ...)
     #panel(eta.mat, sqres, col = rep(col, each = n), cex = 1, cex.lab = 1, cex.axis = 1, lwd = 1)